YOUR_TEAM_NAME/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.3",
    "paho-mqtt>=2.1.0",
    "scipy>=1.16.2",
]
